move trainer builder file
