+++ examples/nlp/language_modeling/megatron_bert_pretraining.py
@@ -13,10 +13,10 @@
 # limitations under the License.
 
 import torch.multiprocessing as mp
-from megatron_trainer_builder import MegatronBertTrainerBuilder
 from omegaconf.omegaconf import OmegaConf, open_dict
 
 from nemo.collections.nlp.models.language_modeling.megatron_bert_model import MegatronBertModel
+from nemo.collections.nlp.parts.megatron_trainer_builder import MegatronBertTrainerBuilder
 from nemo.core.config import hydra_runner
 from nemo.utils import logging
 from nemo.utils.exp_manager import exp_manager
