+++ examples/nlp/language_modeling/megatron_t5_pretraining.py
@@ -13,10 +13,10 @@
 # limitations under the License.
 
 
-from megatron_trainer_builder import MegatronT5TrainerBuilder
 from omegaconf.omegaconf import OmegaConf, open_dict
 
 from nemo.collections.nlp.models.language_modeling.megatron_t5_model import MegatronT5Model
+from nemo.collections.nlp.parts.megatron_trainer_builder import MegatronT5TrainerBuilder
 from nemo.core.config import hydra_runner
 from nemo.utils import logging
 from nemo.utils.exp_manager import exp_manager
